Trigger tracking not at program end or termination
Triggering the tracking at the end of the program causes
unintended behavior. For example, it produces an output,
even if the tracking was configured to only produce outputs
up to a given maximal point in time.
Also in combination with APESmate the triggering at the
end of the program results in an output at every step of
the overall problem solution.